第二次提交

+++ include/webcore/data/ImageStruct.h
@@ -6,7 +6,7 @@ struct ImageStruct
     int width; // 图像宽度
     int height;// 图像高度
     int channels;// 图像通道数
-    int size;   // 图像大小
+    int64_t size;   // 图像大小
     // 使用智能指针包装数据
     std::unique_ptr<unsigned char[]> data;
     // 构造函数和析构函数
